requirements.txt
# requirements.txt
aiogram==3.22.0
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.16.4
python-dotenv==1.1.1
aiohttp==3.12.15
SQLAlchemy==2.0.43
Babel==2.17.0
python-i18n==0.3.9
loguru==0.7.3
httpx==0.28.1
tqdm==4.67.1
rich==14.1.0
pydantic-settings==2.10.1
aiosqlite==0.21.0
fluent-compiler==1.1
fluent-syntax==0.19.0
selenium==4.35.0
webdriver-manager==4.0.2
Pillow==11.3.0
pytesseract==0.3.13
opencv-python==4.12.0.88
numpy==2.2.6
